requirements.txt
flask
chatterbot2
chatterbot-corpus

